+++ qfitswidget/qt/fitswidget.py
