+++ qfitswidget/qt/fitswidget.ui
