+++ qfitswidget/qt/resources.py
