+++ qfitswidget/qt/resources.rc
